window-mixin: Drop the outline node
In GTK4 the outline can be done with just CSS, as follows:

window {
  outline-color: alpha(white, .1);
  outline-width: 1px;
  outline-offset: -1px;
  outline-style: solid;
}

With this in mind, we don't need to have an extra css node and can drop it.

Still keep the contents node, we may need it later, for example for a toast
system.

Fixes https://gitlab.gnome.org/exalm/libadwaita/-/issues/20